% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{Build_df}
\alias{Build_df}
\title{build one comprehensive dataset as a time series (in 2 formats)}
\usage{
Build_df(infile1, infile2, infile3)
}
\arguments{
\item{infile3}{}
}
\value{
2 time.series objects, df (class: xts, zoo - from xts) and df_ts (class: mts, ts, matrix - from ts)
}
\description{
build one comprehensive dataset as a time series (in 2 formats)
}
