% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{clean_data}
\alias{clean_data}
\title{cleans data (removes all columns and lines that are irrelevant)}
\usage{
clean_data(x)
}
\arguments{
\item{infile}{the data.frame}
}
\value{
a data.dable containing only date and Realized Variance based on 5min intervals
}
\description{
cleans data (removes all columns and lines that are irrelevant)
}
