% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{Clean_Vix1}
\alias{Clean_Vix1}
\title{Cleans the VIX data with date format YYYY-MM-DD}
\usage{
Clean_Vix1(infile)
}
\arguments{
\item{infile}{}
}
\value{
a data.frame (tbl_df, tbl, data.frame) with Date (class: Date) and VIX.Close (class: numeric), VIX.Close keeps 2 decimals
}
\description{
Cleans the VIX data with date format YYYY-MM-DD
}
\examples{
Clean_Vix1(Vix1_raw)
}
