% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{Clean_Vix2}
\alias{Clean_Vix2}
\title{Cleans the VIX data with date format D/M/YYYY}
\usage{
Clean_Vix2(infile)
}
\arguments{
\item{infile}{}
}
\value{
a data.frame with Date (class: Date) and VIX.Close (class: numeric), VIX.Close keeps 2 decimals
}
\description{
Cleans the VIX data with date format D/M/YYYY
}
\examples{
Clean_Vix2(Vix2_raw)
}
