% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{Clean_Vol}
\alias{Clean_Vol}
\title{Cleans the Volatility data}
\usage{
Clean_Vol(infile)
}
\arguments{
\item{infile}{}
}
\value{
a data.frame with a Date (class: Date) and a Volatility (Volatility, class: numeric)
}
\description{
Cleans the Volatility data
}
\examples{
Clean_Vol(Vol_raw)
}
