% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Loads data from csv files}
\usage{
load_data(infile)
}
\arguments{
\item{infile}{the datafile}
}
\value{
a data.frame df of the file
}
\description{
Loads data from csv files
}
\examples{
infile <- "data-raw/sandp500.csv"
load_data(infile)
}
